fjerne klasse PersonNavn fra common
