+++ apps/etterlatte-pdltjenester/src/main/kotlin/personweb/dto/PersonNavn.kt
@@ -1,0 +1,10 @@
+package personweb.dto
+
+import no.nav.etterlatte.libs.common.person.Folkeregisteridentifikator
+
+data class PersonNavn(
+    val fornavn: String,
+    val mellomnavn: String? = null,
+    val etternavn: String,
+    val foedselsnummer: Folkeregisteridentifikator,
+)
