+++ libs/saksbehandling-common/src/main/kotlin/person/Person.kt
@@ -43,13 +43,6 @@ data class Person(
     var vergemaalEllerFremtidsfullmakt: List<VergemaalEllerFremtidsfullmakt>?,
 )
 
-data class PersonNavn(
-    val fornavn: String,
-    val mellomnavn: String? = null,
-    val etternavn: String,
-    val foedselsnummer: Folkeregisteridentifikator,
-)
-
 data class Statsborgerskap(
     val land: String,
     val gyldigFraOgMed: LocalDate?,
